address review and rename example directories and some more alignments
