+++ example-chain-opensearch/opensearch_chain/constants/constants.py
