+++ example-chain-opensearch/opensearch_chain/services/opensearch_service.py
