+++ release-modules
@@ -1,4 +1,4 @@
-basic-actor-chain-example
+example-chain-basic-actor
 virtualizer-chain-example
 opensearch-chain-example
-memory-chain-example
+example-chain-chat-memory
